refactor: use shared package for chakra-ui/utils
