+++ packages/c-alert/package.json
@@ -23,16 +23,16 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-utils": "^0.0.1-alpha.0",
-    "@chakra-ui/c-icon": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
+    "@chakra-ui/vue-utils": "0.0.1-alpha.0",
+    "@chakra-ui/c-icon": "0.0.1-alpha.0"
   },
   "devDependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
     "vue": "^3.0.5"
   },
   "peerDependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
     "vue": "^3.0.5"
   }
 }
