+++ packages/c-badge/src/badge.ts
@@ -5,7 +5,7 @@ import {
   ThemingProps,
   useStyleConfig,
 } from '@chakra-ui/vue-system'
-import { filterUndefined } from '@chakra-ui/vue-utils'
+import { filterUndefined } from '@chakra-ui/utils'
 
 const CBadge = defineComponent({
   props: {
@@ -19,8 +19,8 @@ const CBadge = defineComponent({
     styleConfig: String as PropType<ThemingProps['styleConfig']>,
   },
   setup(props, { slots, attrs }) {
-    const themingProps = computed(() =>
-      filterUndefined<ThemingProps>({
+    const themingProps = computed<ThemingProps>(() =>
+      filterUndefined({
         colorScheme: props.colorScheme,
         variant: props.variant,
         size: props.size,
