+++ packages/c-reset/package.json
@@ -36,14 +36,14 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-system": "0.0.1-alpha.0"
   },
   "peerDependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
     "vue": ">=3.0.5"
   },
   "devDependencies": {
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
     "vue": ">=3.0.5"
   }
 }
