+++ packages/c-spinner/package.json
@@ -34,20 +34,20 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/c-visually-hidden": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0"
+    "@chakra-ui/c-visually-hidden": "0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0"
   },
   "devDependencies": {
-    "@chakra-ui/c-visually-hidden": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0",
+    "@chakra-ui/c-visually-hidden": "0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0",
     "vue": "^3.0.5"
   },
   "peerDependencies": {
-    "@chakra-ui/c-visually-hidden": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-system": "^0.0.1-alpha.0",
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0",
+    "@chakra-ui/c-visually-hidden": "0.0.1-alpha.0",
+    "@chakra-ui/vue-system": "0.0.1-alpha.0",
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0",
     "vue": "^3.0.5"
   }
 }
