+++ packages/c-theme-provider/package.json
@@ -29,12 +29,12 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0"
   },
   "peerDependencies": {
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0"
   },
   "devDependencies": {
-    "@chakra-ui/vue-theme": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-theme": "0.0.1-alpha.0"
   }
 }
