+++ packages/core/src/parse-icons.ts
@@ -1,4 +1,4 @@
-import { mergeWith } from '@chakra-ui/vue-utils'
+import { mergeWith } from '@chakra-ui/utils'
 
 interface InternalIcon {
   [key: string]:
