+++ packages/system/package.json
@@ -20,14 +20,16 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/c-color-mode": "^0.0.1-alpha.0",
+    "@chakra-ui/c-color-mode": "0.0.1-alpha.0",
     "@chakra-ui/styled-system": "^1.9.0",
-    "@chakra-ui/vue-utils": "^0.0.1-alpha.0",
+    "@chakra-ui/utils": "1.4.0",
+    "@chakra-ui/vue-utils": "0.0.1-alpha.0",
     "@emotion/css": "^11.1.3",
     "lodash.camelcase": "^4.3.0",
     "lodash.kebabcase": "^4.1.1"
   },
   "peerDependencies": {
+    "@chakra-ui/utils": "1.4.0",
     "vue": ">=3.0.5"
   }
 }
