+++ packages/system/src/composables/use-style-config.ts
@@ -1,7 +1,7 @@
 import { computed, ComputedRef } from 'vue'
 import { SystemStyleObject } from '@chakra-ui/styled-system'
 import { ThemingProps } from '../system.types'
-import { filterUndefined, get, mergeWith, runIfFn } from '@chakra-ui/vue-utils'
+import { filterUndefined, get, mergeWith, runIfFn } from '@chakra-ui/utils'
 import { useChakra } from './use-chakra'
 
 export function useStyleConfig(
