+++ packages/system/src/system.utils.ts
@@ -1,4 +1,4 @@
-import { UnionStringArray } from '@chakra-ui/vue-utils'
+import { UnionStringArray } from '@chakra-ui/utils'
 import { keyframes, injectGlobal } from '@emotion/css'
 
 /**
