+++ packages/theme/package.json
@@ -23,6 +23,6 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@chakra-ui/vue-theme-tools": "^0.0.1-alpha.0"
+    "@chakra-ui/vue-theme-tools": "0.0.1-alpha.0"
   }
 }
