+++ packages/utils/package.json
@@ -20,10 +20,12 @@
     "watch:types": "cross-env tsc --emitDeclarationOnly --declaration --declarationDir dist/types --watch --incremental"
   },
   "dependencies": {
-    "@types/lodash.mergewith": "4.6.6",
-    "@types/object-assign": "4.0.30",
     "css-box-model": "1.2.1",
     "lodash.mergewith": "4.6.2",
     "object-assign": "4.1.1"
+  },
+  "devDependencies": {
+    "@types/lodash.mergewith": "4.6.6",
+    "@types/object-assign": "4.0.30"
   }
 }
