+++ packages/utils/src/layout.ts
@@ -1,0 +1,9 @@
+export function orient(options: {
+  orientation?: 'vertical' | 'horizontal'
+  vertical: any
+  horizontal: any
+}) {
+  const { orientation, vertical, horizontal } = options
+  if (!orientation) return {}
+  return orientation === 'vertical' ? vertical : horizontal
+}
